requirements.txt
anthropic==0.88.0
fastapi==0.135.3
httpx==0.28.1
psycopg[binary]==3.3.3
psycopg-pool==3.3.0
python-dotenv==1.2.2
uvicorn==0.42.0
